Add cacheItemCallback to CacheService
Callback is used to determine if item can be cached. Result from
function call is passed to callback and if it returns true, it is
cached. Otherwise not.

It is useful for situations when certain object can be cached when it
is in some specific state (for example PackageMetadata can be cached
when scanStatus is "COMPLETE").