Fix caching for Object types
Object types cannot be converted to string by .toString() method,
because it leads to "[object Object]" result and it is the same for
all objects. This issue is fixed by using JSON.stringify. It gives
different results for different objects, but the same results when
object has the same parameters and only instance is different.