Remove excessive logging
It is OK to use console.log for quick debugging while development,
but I think we should remove it when feature is finished. We had
quite a lot of logging in code and console became quite messy.

I removed most of console.log statements and kept there only warn or
error ones, because these signal something wrong is going on and
should not get triggered in common use.