Persist AnalysisContext configuration
Each project should have defaultAnalysisContext. This context is used
for configuration. It is never used for execution. Execution creates
a clone of that AnalysisContext.

There is fallback solution if defaultAnalysisContext doesn't exist on
MigrationProject (even though this should never happen!). Also
analysisContext should never be updated, when it is used for execution.