Wrong color for the inner-frame (and too low contrast) refer to the UI flow
The issue with the background color is that we use quite dark one for the page background. Almost the same as the mockup wizard bg - F3F3F3.
The UI mocks have sometimes gray page bg, sometimes white. I am keeping gray and using darker #e8e8e8 for this.