808 - wizard fixes (#370)
* Show RegisterApp form errors as global errors

* Fix RegisterApp form submit for server path

When there are some applications already uploaded and user is on
server path tab, assume form as valid even when path is not set. But
do not send a request and proceed with navigation right away.

* Update AnalysisContextForm loading status for packages

- rewording and positioning

* Make project title help text always the same color

Project title help text should not change to red when there is
validation error.